test_requirements.txt
pytest==7.1.2
pytest-playwright==0.3.0